<Feature, Optimisation> Pathfinding, Brands
Improved pathfinding and related functionality.
Moved FindCellToAttackPlayer to enemy, now simply finds a route to the
player and stops when it is within range- reduces the calls to
findcellsnearme.
Cover is found by using point in polygon on the player light- consider
anywhere in darkness to be cover.
Improved move to player functionality, reduced the number of calls
enemies make to the pathfinding system.
Removed class-based attribute caps, classes are more clearly defined at
the beginning of play.
Added overcharge ui.
Skill UI only displays when skill is unlocked.
Added other brand related features.
Added decay particle system.
Skills require adrenaline.
Adrenaline recharges over time.

+++ Night Unity Files/Assets/Scripts/Game/Combat/Player/Skill.cs
@@ -35,7 +35,7 @@ namespace Game.Combat.Player
             s._skillValue = value;
         }
 
-        public float Cooldown()
+        public int Cooldown()
         {
             return _skillValue.Cooldown;
         }
